+++ CHANGELOG.md
@@ -24,6 +24,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Implement Catalog service for Data Management API (#1195)
 * Add strict body validation for REST endpoints (#1128)
 * Dependency injection using factory/provider methods (#1056)
+* Provisioned resource information in Data Management API (#1221)
 
 #### Changed